Move the merge into core.py and run it inside the LLK container
Review feedback: the merge belongs in helpers.perf.core, not its own module.
Removes helpers/perf/merge.py and appends its functions to core.py, keeping
your trimmed docstrings; `main` becomes `merge_main`, since a bare `main` in a
1200-line library module says nothing. Also trims the remaining long comments
this branch added in core.py, test_config.py and test_perf_report_hw_free.py.

Moving it fixed a real defect. The merge step ran on ubuntu-latest with only
pyarrow installed, but any `helpers.*` import pulls
helpers/__init__.py -> ttexalens, so the step would have failed with
ModuleNotFoundError on its first real run -- the local tests only pass because
they stub ttexalens. The job now runs in the LLK CI image, which already has
the environment core.py needs, so the setup-python and pip steps are gone.

One thing to watch on the first dispatch: the job now needs `sftp` inside that
image. If openssh-client is absent there, the fallback is one line -- drop the
container and install the test requirements instead.
