+++ .github/workflows/llk-perf-impl.yaml
@@ -201,9 +201,14 @@ jobs:
 
   upload-perf-data:
     name: "Upload perf data to the warehouse"
-    needs: [llk-perf-exec]
+    needs: [build-images, llk-perf-exec]
     if: ${{ always() && inputs.upload-to-warehouse }}
     runs-on: ubuntu-latest
+    # The merge lives in helpers.perf.core, whose imports reach the device
+    # libraries, so it needs the LLK python env. The test image has it.
+    container:
+      image: ${{ needs.build-images.outputs.llk-ci-tag }}
+      options: "--rm"
     steps:
       - name: Checkout the repository
         uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
@@ -215,22 +220,16 @@ jobs:
         with:
           pattern: perf-data-*
           path: perf-data
-      - name: Set up Python
-        uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6.2.0
-        with:
-          python-version: "3.10"
-      - name: Install the merge dependency
-        run: pip install "pyarrow==24.0.0"
       - name: Merge the run's shards into one file per architecture
         working-directory: ${{ env.LLK_PATH }}/tests/python_tests
         env:
           PYTHONPATH: .
+          CHIP_ARCH: wormhole
         run: |
-          python3 -m helpers.perf.merge \
+          python3 -m helpers.perf.core \
             --in-dir "$GITHUB_WORKSPACE/perf-data" \
             --out-dir "$GITHUB_WORKSPACE/generated/llk_perf"
           ls -l "$GITHUB_WORKSPACE/generated/llk_perf"
-
       - name: Upload perf data
         uses: ./.github/actions/upload-data-via-sftp
         with:
